Bayesian inference of a brift diffusion model

Bayesian inference of a brift diffusion model

ddm_stan(formula, data, tau = 0.1, alpha_ub = 5, sigma = 2, ...)

Arguments

formula
A formula of the kind y + z ~ 1

data
A data frame with response times and response choices

tau
The non-decision time

alpha_ub
The upper bound on the uniform prior on alpha

sigma
The standard deviation of the normal prior on delta

...
Arguments passed to rstan::sampling (e.g. chains, cores)

Value

A stan model with MCMC samples

Examples

set.seed(10101)
sims_df <- simulate_ddm(1000, b = 0.75, a = 2, v = 0.5) %>%
    dplyr::mutate(time = time + 0.1)
ddm_stan(time + choice ~ 1, data = sims_df, cores = 4, refresh = 0)
#> Inference for Stan model: ddm.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>          mean se_mean   sd    2.5%     25%     50%     75%   97.5% n_eff Rhat
#> alpha    1.99    0.00 0.04    1.92    1.97    1.99    2.02    2.08  2021    1
#> beta     0.74    0.00 0.01    0.73    0.74    0.74    0.75    0.76  1755    1
#> delta    0.47    0.00 0.05    0.38    0.43    0.47    0.50    0.56  1759    1
#> lp__  -705.94    0.03 1.28 -709.17 -706.50 -705.62 -705.03 -704.50  1784    1
#> 
#> Samples were drawn using NUTS(diag_e) at Fri Dec 10 14:29:24 2021.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).